+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 6.22.0 (December 19, 2022). Tested on Artifactory 7.47.12
+
+IMPROVEMENTS:
+
+* Added new datasource: datasource/datasource_artifactory_group.go
+  Issue: [#548](https://github.com/jfrog/terraform-provider-artifactory/issues/548)
+  PR: [#607](https://github.com/jfrog/terraform-provider-artifactory/pull/607)
+
 ## 6.21.8 (December 15, 2022). Tested on Artifactory 7.47.12
 
 IMPROVEMENTS: